Move concatenate operation from 'iris.experimental' into a CubeList method.
